requirements.txt
aiohttp>=3.8.1
tqdm>=4.64.0
pydantic>=1.9.1
py-xmltv==1.0.8
Pillow>=9.1.1
diskcache>=5.4.0
furl>=2.1.3
lxml>=4.9.0